refactor: separate auto range class
