+++ src/widget/auto_range_lerp.css
